rename content creator files
